chore: Add docker-compose file for PostgreSQL and pgAdmin
This commit adds a docker-compose file for setting up PostgreSQL and pgAdmin containers. The PostgreSQL container is based on version 15 of the postgres image and is configured with the default username and password. The pgAdmin container is based on the dpage/pgadmin4 image and is configured with the default email and password. The docker-compose file also includes a volume for persisting the PostgreSQL data and a network for connecting the containers.

The commit also includes the servers.json file, which contains the configuration for connecting to the PostgreSQL server in pgAdmin.

The purpose of this commit is to provide an easy way to set up a development environment with PostgreSQL and pgAdmin.

```