feat: Add Google authentication fields to User model
This commit adds the `googleId` and `profilePicture` fields to the `User` model in the Prisma schema. The `googleId` field is marked as unique, allowing users to authenticate using their Google account. The `profilePicture` field is added to store the URL of the user's profile picture.

The purpose of this commit is to enable Google authentication and provide a way to store and retrieve the user's profile picture.