chore: Update docker-compose.api-nest.yml and package.json
This commit updates the `docker-compose.api-nest.yml` file to use the `${PORT}` environment variable instead of `${SERVER_PORT}` for the exposed port. This change aligns with the updated configuration.

Additionally, the `package.json` file is updated to include the latest version of `webpack` as a dependency. This ensures compatibility with the latest features and improvements.

The purpose of this commit is to update the configuration and dependencies to maintain a stable and up-to-date development environment.