chore: Update prisma binaryTargets in schema.prisma
This commit updates the `binaryTargets` property in the `generator client` section of the `schema.prisma` file. The new value includes the `debian-openssl-3.0.x` target, in addition to the existing targets for `linux-musl`, `windows`, and `darwin-arm64`.

The purpose of this commit is to ensure compatibility with the `debian-openssl-3.0.x` target, allowing the generated Prisma client to be used in Debian-based environments with OpenSSL version 3.0.x.